Added ball class
Needed to add the basics of the ball class for use in the bot. The constructor takes the id for the ball aka ball number and the x and y quards of the position. these values are saved to a class variable id and position where position is a touple of (x, y)